+++ src/colors/syntax/ts.colors.ts
@@ -11,12 +11,19 @@ const tsColors: SyntaxColors = (tokens) => {
       settings: tokens.class.name
     },
 
-    /**
-     * Keywords
-     */
     {
-      name: 'JSON object',
-      scope: 'support.constant.json',
+      name: 'Arrow function. =>',
+      scope: 'storage.type.function.arrow.ts',
+      settings: tokens.javascript.arrowFunction
+    },
+    {
+      name: 'Other operators',
+      scope: [
+        'keyword.operator.new.ts',
+        'keyword.operator.in.ts',
+        'keyword.operator.of.ts',
+        'keyword.operator.typeof.ts'
+      ],
       settings: tokens.keyword.other
     }
   ];